Final lab4 with checked request.http files

+++ person-management/src/main/java/pl/edu/pg/aui/personmanagement/service/impl/PersonDefaultService.java
@@ -1,6 +1,7 @@
 package pl.edu.pg.aui.personmanagement.service.impl;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.web.client.RestTemplate;
@@ -20,10 +21,14 @@ public class PersonDefaultService implements PersonService {
 
     private final RestTemplate restTemplate;
 
+    private final String personManagementUrl;
+
+
     @Autowired
-    public PersonDefaultService(PersonRepository personRepository, RestTemplate restTemplate) {
+    public PersonDefaultService(PersonRepository personRepository, RestTemplate restTemplate, @Value("${person.management.url}") String personManagementUrl) {
         this.personRepository = personRepository;
         this.restTemplate = restTemplate;
+        this.personManagementUrl = personManagementUrl;
     }
 
     @Override
@@ -55,16 +60,17 @@ public class PersonDefaultService implements PersonService {
     public void delete(UUID id) {
         Person person = personRepository.findById(id)
                 .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Person not found"));
-        String url = "${person.management.url}" + id + "/cars";
+        String url = personManagementUrl + "/" + id + "/cars";
         restTemplate.delete(url);
         personRepository.delete(person);
     }
 
     @Override
     public void delete(Person person) {
-        Person tmp = personRepository.findById(person.getId())
-                .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Person not found"));
-        String url = "${person.management.url}" + person.getId() + "/cars";
+        if(!personRepository.existsById(person.getId())) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Person not found");
+        }
+        String url = personManagementUrl + "/" + person.getId() + "/cars";
         restTemplate.delete(url);
         personRepository.delete(person);
     }
